-Split and reorginized files -Added first summary generation code
